+++ globalConfig.json
@@ -1,1 +1,1 @@
-{"mongoUri":"mongodb://127.0.0.1:46087/2c730213-4457-41f4-ab37-4d30326595f6?"}
+{"mongoUri":"mongodb://127.0.0.1:45177/7664bdbf-b85f-4619-ac33-54c852019ce5?"}
