+++ src/presentation/controllers/signup/signup.spec.ts
@@ -7,7 +7,7 @@ import {
   HttpRequest,
   Validation
 } from './signup-protocols'
-import { ok, serverError, badRequest } from '../../helpers/http-helper'
+import { ok, serverError, badRequest } from '../../helpers/http/http-helper'
 
 const makeAddAccount = (): AddAccount => {
   class AddAccountStub implements AddAccount {
