+++ src/presentation/controllers/signup/signup.ts
@@ -5,7 +5,7 @@ import {
   AddAccount,
   Validation
 } from './signup-protocols'
-import { badRequest, serverError, ok } from '../../helpers/http-helper'
+import { badRequest, serverError, ok } from '../../helpers/http/http-helper'
 
 export class SignUpController implements Controller {
   private readonly addAccount: AddAccount
